+++ src/Source/DatabaseSource.php
@@ -2,12 +2,10 @@
 
 namespace Rougin\Transcribe\Source;
 
-use PDO;
-
 /**
  * Database Source
  *
- * Retrieves a list of words via PHP Database Object.
+ * Returns an array of words using the PHP Database Object (PDO).
  *
  * @package Transcribe
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
@@ -25,37 +23,59 @@ class DatabaseSource implements SourceInterface
     protected $table;
 
     /**
-     * @param PDO   $pdo
+     * @var array
+     */
+    protected $words = array();
+
+    /**
+     * Initializes the source instance.
+     *
+     * @param \PDO  $pdo
      * @param array $table
      */
-    public function __construct(PDO $pdo, array $table)
+    public function __construct(\PDO $pdo, array $table)
     {
-        $this->pdo   = $pdo;
+        $this->pdo = $pdo;
+
         $this->table = $table;
     }
 
     /**
-     * Returns a list of words.
+     * Returns an array of words.
+     * NOTE: To be removed in v1.0.0. Use "words" instead.
      *
      * @return array
      */
     public function getWords()
     {
-        $words = [];
+        return $this->words();
+    }
 
-        $language    = $this->table['language'];
-        $text        = $this->table['text'];
-        $translation = $this->table['translation'];
-
+    /**
+     * Returns an array of words.
+     *
+     * @return array
+     */
+    public function words()
+    {
         $query = 'SELECT * FROM ' . $this->table['name'];
+
         $table = $this->pdo->prepare($query);
 
         $table->execute();
 
-        foreach ($table->fetchAll(PDO::FETCH_ASSOC) as $row) {
-            $words[$row[$language]][$row[$text]] = $row[$translation];
+        foreach ($table->fetchAll(\PDO::FETCH_ASSOC) as $row) {
+            $group = $row[$this->table['language']];
+
+            isset($this->words[$group]) || $this->words[$group] = array();
+
+            $text = (string) $row[$this->table['text']];
+
+            $translation = $row[$this->table['translation']];
+
+            $this->words[$group][$text] = $translation;
         }
 
-        return $words;
+        return $this->words;
     }
 }