+++ src/Transcribe.php
@@ -8,7 +8,7 @@ use Rougin\Transcribe\Source\SourceInterface;
 /**
  * Transcribe
  *
- * Yet another language library for PHP.
+ * An easy-to-use localization library for PHP.
  *
  * @package Transcribe
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
@@ -16,47 +16,99 @@ use Rougin\Transcribe\Source\SourceInterface;
 class Transcribe
 {
     /**
-     * @var \Rougin\Transcribe\Source\SourceInterface
-     */
-    protected $source;
-
-    /**
-     * List of words.
-     *
      * @var array
      */
-    protected $vocabulary = [];
+    protected $dotified = array();
 
     /**
-     * @param SourceInterface $source
+     * @var array
+     */
+    protected $words = array();
+
+    /**
+     * Initializes the Transcribe instance.
+     *
+     * @param \Rougin\Transcribe\Source\SourceInterface $source
      */
     public function __construct(SourceInterface $source)
     {
-        $this->source     = $source;
-        $this->vocabulary = $source->getWords();
+        // Note: Use $source->words() in v1.0.0.
+        $this->words = $source->getWords();
+
+        $this->dotified = $this->dotify($this->words);
     }
 
     /**
-     * Gets the specified text from the list of defined words in the file.
+     * Returns all words stored from the source.
      *
-     * @param  string $string
+     * @return array
+     */
+    public function all()
+    {
+        return $this->words;
+    }
+
+    /**
+     * Returns the specified text from an array of words.
+     * NOTE: To be removed in v1.0.0. Use "get" instead.
+     *
+     * @param  string $text
      * @return string
      */
-    public function getText($string = '')
+    public function get($text)
     {
-        $multiArray = new MultiArray($this->vocabulary);
-        $result     = $multiArray->get($string);
+        if (! isset($this->dotified[$text])) {
+            $this->dotified[$text] = $text;
+        }
 
-        return (is_string($result)) ? $result : $string;
+        return $this->dotified[$text];
     }
 
     /**
-     * Returns all words stored from the vocabulary.
+     * Returns the specified text from an array of words.
+     * NOTE: To be removed in v1.0.0. Use "get" instead.
+     *
+     * @param  string $text
+     * @return string
+     */
+    public function getText($text)
+    {
+        return $this->get($text);
+    }
+
+    /**
+     * Returns all words stored from the source.
+     * NOTE: To be removed in v1.0.0. Use "all" instead.
      *
      * @return array
      */
     public function getVocabulary()
     {
-        return $this->vocabulary;
+        return $this->all();
+    }
+
+    /**
+     * Converts the data into dot notation values.
+     *
+     * @param  array  $data
+     * @param  array  $result
+     * @param  string $key
+     * @return array
+     */
+    protected function dotify(array $data, $result = array(), $key = '')
+    {
+        foreach ((array) $data as $name => $value) {
+            if (is_array($value) === true) {
+                $new = (string) $key . $name . '.';
+
+                $array = $this->dotify($value, $result, $new);
+
+                $result = array_merge($result, $array);
+            }
+
+            is_array($value) || $result[$key . $name] = $value;
+        }
+
+        return $result;
     }
 }