+++ tests/Fixture/Languages/Locales/fil_PH.php
@@ -1,7 +1,0 @@
-<?php
-
-return [
-    'name'     => 'pangalan',
-    'language' => 'linguahe',
-    'school'   => 'paaralan',
-];