+++ tests/Fixture/Locales/fil_PH.php
@@ -1,0 +1,9 @@
+<?php
+
+$texts = array();
+
+$texts['language'] = 'linguahe';
+$texts['name'] = 'pangalan';
+$texts['school'] = 'paaralan';
+
+return $texts;