+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 # Changelog
 
-All Notable changes to `Transcribe` will be documented in this file
+All notable changes to `Transcribe` will be documented in this file.
+
+## [0.4.0](https://github.com/rougin/transcribe/compare/v0.3.1...v0.4.0) - Unreleased
+
+### Added
+- `SourceCollection::add`
+- `SourceInterface::words`
+- `Transcribe::all`
+- `Transcribe::get`
+
+### Changed
+- Minimum version of PHP to `v5.3.0`
+
+### Deprecated
+- `SourceCollection::addSource`
+- `SourceInterface::getWords`
+- `Transcribe::getText`
+- `Transcribe::getVocabulary`
+
+### Removed
+- `CONTRIBUTING.md`
+- `tebru/multi-array` package
 
 ## [0.3.1](https://github.com/rougin/transcribe/compare/v0.3.0...v0.3.1) - 2016-09-11
 
@@ -10,7 +31,10 @@ All Notable changes to `Transcribe` will be documented in this file
 ## [0.3.0](https://github.com/rougin/transcribe/compare/v0.2.2...v0.3.0) - 2016-07-29
 
 ### Added
-- `SourceCollection`
+- `Source\SourceCollection`
+
+### Deprecated
+- `Source\MultipleSource`
 
 ## [0.2.2](https://github.com/rougin/transcribe/compare/v0.2.1...v0.2.2) - 2016-03-12
 
@@ -20,7 +44,7 @@ All Notable changes to `Transcribe` will be documented in this file
 ## [0.2.1](https://github.com/rougin/transcribe/compare/v0.2.0...v0.2.1) - 2015-09-30
 
 ### Added
-- [MultiArray](https://github.com/tebru/multi-array) for handling multidimensional arrays
+- `tebru/multi-array` package for handling multidimensional arrays
 
 ## [0.2.0](https://github.com/rougin/transcribe/compare/v0.1.1...v0.2.0) - 2015-09-27
 
@@ -28,8 +52,7 @@ All Notable changes to `Transcribe` will be documented in this file
 - [Source](https://github.com/rougin/transcribe/tree/master/src/Source) directory for handling data from different sources
 
 ### Fixed
-- Code structure
-    - Conformed to [SOLID](https://en.wikipedia.org/wiki/SOLID_(object-oriented_design))-based design
+- Conformed code structure to [SOLID](https://en.wikipedia.org/wiki/SOLID_(object-oriented_design))-based design
 
 ### Deprecated
 - Complex constructors in instantiating the library